test: extract unit tests into own file for ChannelSubscriptions functionality
