+++ pushpin-security-core/src/test/kotlin/io/github/mpecan/pmt/security/model/ChannelSubscriptionTest.kt
@@ -50,72 +50,3 @@ class ChannelSubscriptionTest {
             .containsExactly("sales", "revenue")
     }
 }
-
-class ChannelSubscriptionsTest {
-    @Test
-    fun `should check if subscription is allowed`() {
-        val subscriptions =
-            ChannelSubscriptions(
-                principal = "user123",
-                subscriptions =
-                    listOf(
-                        ChannelSubscription("news", allowed = true),
-                        ChannelSubscription("admin", allowed = false),
-                        ChannelSubscription("user.123", allowed = true),
-                    ),
-                defaultAllow = false,
-            )
-
-        assertThat(subscriptions.canSubscribe("news")).isTrue()
-        assertThat(subscriptions.canSubscribe("admin")).isFalse()
-        assertThat(subscriptions.canSubscribe("user.123")).isTrue()
-        assertThat(subscriptions.canSubscribe("unknown")).isFalse()
-    }
-
-    @Test
-    fun `should respect default allow policy`() {
-        val subscriptions =
-            ChannelSubscriptions(
-                principal = "user123",
-                subscriptions =
-                    listOf(
-                        ChannelSubscription("blocked", allowed = false),
-                    ),
-                defaultAllow = true,
-            )
-
-        assertThat(subscriptions.canSubscribe("blocked")).isFalse()
-        assertThat(subscriptions.canSubscribe("any-other-channel")).isTrue()
-    }
-
-    @Test
-    fun `should get all allowed channels`() {
-        val subscriptions =
-            ChannelSubscriptions(
-                principal = "user123",
-                subscriptions =
-                    listOf(
-                        ChannelSubscription("news", allowed = true),
-                        ChannelSubscription("admin", allowed = false),
-                        ChannelSubscription("user.123", allowed = true),
-                        ChannelSubscription("private", allowed = false),
-                    ),
-            )
-
-        assertThat(subscriptions.getAllowedChannels())
-            .containsExactlyInAnyOrder("news", "user.123")
-    }
-
-    @Test
-    fun `should handle empty subscriptions`() {
-        val subscriptions =
-            ChannelSubscriptions(
-                principal = "user123",
-                subscriptions = emptyList(),
-                defaultAllow = false,
-            )
-
-        assertThat(subscriptions.getAllowedChannels()).isEmpty()
-        assertThat(subscriptions.canSubscribe("any-channel")).isFalse()
-    }
-}
